NFDIV-1909 - Update IrretrievablyBroken question API mapping
As the field for marriageIrretrievablyBroken has changed on the back end
from a YesOrNo field to a Set, this PR will update the mapping of the field.

Update definition.ts which was generated by the new API version.

Add new mapping to and from API for applicant1ScreenHasUnionBroken.

Fix issues found due to update in definition.ts file.

Update unit tests.